Update price type from String to Double.

+++ app/src/main/java/com/example/futurestore/CartShowProductActivity.kt
@@ -49,7 +49,7 @@ class CartShowProductActivity : AppCompatActivity() {
                 ProductInformation(
                     productName,
                     productCategory,
-                    productPrice,
+                    productPrice.toDouble(),
                     1,
                     productImageLink,
                     productDescription
